refactor: split vsock host file helpers
